requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
fastapi==0.75.2
FastAPI-SQLAlchemy==0.2.1
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.2.0
uvicorn==0.17.6
yarl==1.7.2
